gs-overview-page: Support a second category flowbox for addons
This will be for less important categories, such as add-on categories,
which won’t have custom styling.

Currently, the addons are mixed into the Create and Work categories, and
this doesn’t work very well — the addons are hard to find due to there
being a small number of them mixed in with a large number of other apps.
A lot of fonts start with ‘a’ so end up monopolising the top of the
alphabetic list of software. Instructing someone to open the ‘Work’
category to browse for hardware drivers is confusing.

A following commit will rearrange the category map to separate the
addons out again.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1384