gs-desktop-data: Split addons out into separate categories
See the previous commit. It didn’t work very well for the addons to be
mixed in with other software, so separate them out into their own
categories again. These will now be listed separately at the bottom of
the overview page, as per the previous commit.

We may want to iterate on this design more in future, but it will do for
gnome-software 41, given the time pressure.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1384