Rework event handling
On GTK4, GtkWidget doesn't have the *-event signals anymore. Events must
be routed through controllers and gestures (e.g. GtkEventController and
GtkGestureClick).

This commit does that.

GsShell needs 2 separate key controllers, one during the capture phase
and the other during bubble, so that it can mimic its current behavior.
It would be ideal to switch to actions later, but for now let's be as
un-disruptive as possible.

Since GtkGestureClick supports setting a button, the 'if (button != 8)'
check can be dropped. To stop event propagation of click gestures, claim
the gesture state.