gs-summary-tile: Use a specialized layout manager to handle allocation
The 'measure' vfunc isn't called when a layout manager is set, and sadly
this is the case for GsAppTile, so we need to replace it with another
layout manager that does what we want to.

Add GsSummaryTileLayout, which is basically just a GtkBinLayout with
an extra step limiting the natural width to a custom value.