Use GsInfoWindow
This homogeneizes and simplifies the many windows implementing this
pattern. This doesn't apply it to GsFirstRunDialog as it doesn't have a close
button, and hence implements a slightly different pattern.

Indentation in UI files is purposefully left broken to make the change
easier to review, and will be fixed in the next commit.