gs-updates-section: Check also dependencies' download size
Since the split of the app download size the code to verify whether
the update is downloaded or not should check both the app itself
and its dependencies. One example are "OS Updates", which is only
a proxy app, gathering system packages to be updated. It does not
have download size on its own.