gs-app-context-bar: Fix re-showing tiles after viewing a non-desktop app
Since !938, some of the context tiles are hidden for non-desktop apps
because they can’t show any information relevant to non-desktop apps.

However, they remained hidden when switching to viewing a desktop app,
which was not the intended behaviour.

This was because `gtk_widget_is_visible()` takes the visibility of the
widget’s parents into account. This meant that once one of the parent
widgets (i.e. the containing box) was hidden, the tiles inside would
never be un-hidden.

Fix that by only taking the visibility of the tiles themselves, and not
their ancestors, into account as was intended when I wrote the code.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1434