packagekit: Ensure PkClient::interactive flag being set
That's required to properly ask or not ask for the root credentials on operations,
which require higher privileges. It's important to set the flag always before the call
to the PkTask/PkClient API, because other thread could change the value while
the corresponding lock was released by the execution thread.

Downstream bug report:
https://bugzilla.redhat.com/show_bug.cgi?id=2003365