gs-plugin-os-release: Use the OS ID rather than NAME in the app ID
When building a `GsApp` to represent the OS, use the `ID` field from
`/etc/os-release` in the app’s ID, rather than the `NAME` field.

On Debian, in particular, the `NAME` field is problematic as it’s
‘Debian GNU/Linux’, with both the space and the `/` not being valid
characters for an ID. They cause these warnings:
```
09:38:54:0415 XbSilo ignoring for OR statement: predicate [text()='debian.www.Debian GNU was unfinished, missing ']'
```

On Fedora, the `GsApp`’s ID is now `org.fedoraproject.fedora-34`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #508