appstream: Ensure XbBuilders are constructed in the right context
This is a follow-up to commit 6fb78f853b8183501fb. While that commit was
complete, it will not be sufficient once libxmlb is modified to store
the `GMainContext` at construction time of an `XbSilo`, which is a
long-term solution to the libxmlb side of the threading problem.

Specifically, `xb_builder_new()` constructs a new `XbSilo`, which may
have a long life. Therefore, `xb_builder_new()` needs to be called in
the correct `GMainContext`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1422