gs-plugin-provenance: Mark fedora-third-party provided apps with GS_APP_QUIRK_DISTRO_SAFE
The fedora-third-party repos are not stored in the 'official-repos'
GSettings key, but they also should not be, because the provenance
quirk is used in the repos dialog to disable some functionality
on such repos. Keeping the both lists (the GSettings key and
the actual fedora-third-party repos list) in sync is also impractical.

The GS_APP_QUIRK_DISTRO_SAFE will be used to show in the safety tile/dialog
that the app is reviewed by the distro, rather than as being from a third-party
repository.