tests: Ensure GTK resources are registered for icon test
The GTK icon resources are needed for the icon test, and they’re
compiled into libgtk. Normally they’re registered when `gtk_init()` is
called, but we can’t call that in the tests as they’re often run
headless.

Call `gtk_init_check()` instead. It will fail to completely set up GTK,
but will not abort the program, and will successfully register the
resources first.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>