gs-appstream: Tidy up another iteration over nodes
This iteration was not buggy and did not leak, but can now be refactored
to use a `for` loop and `node_set_to_next()`. This reduces the number of
lines of code, makes the termination condition for the loop easier to
verify, and factors out the complicated bit of the reference handling.

This commit introduces no functional changes.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>