flatpak: Fix capitalization of boolean value in tests
GLib recognizes "false" but not "False" as false in
g_key_file_get_boolean(). However, because that function returns false
when either the value is false or the function errored out (genius!),
and libflatpak doesn't check for an error in this case, the incorrect
spelling usually has no effect. But with the recent changes in
gs_flatpak_file_to_app_ref(), the "False" spelling leads to a failure of
the unit tests.