gs-extras-page: Break out of loop after removing a row
When searching for duplicate applications, we can assume (by induction)
that there will only be one duplicate instance of an app. Accordingly,
break out of the loop when that instance is removed (before adding the
new version of the app).

This fixes an issue where the next iteration of the loop would call
`gtk_widget_get_next_sibling()` on a widget which had just been
destroyed as a result of being removed from the list.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Fixes: #1473