flatpak: Update remote name for flatpakref unit test
Now that we are using the FlatpakTransaction API instead of
FlatpakInstallation to install flatpakref files, the RuntimeRepo= key is
properly handled and Flatpak uses it to create a remote. In the case of
the unit test for flatpakref installation, the RuntimeRepo file is
called "test.flatpakrepo" so "test" becomes the name of the remote, and
the remote is then used as the origin for the app as well since it has
both the app and the runtime in it. So, update an assertion to account
for the new origin remote name for the test flatpakref file.