gs-installed-page: Change section on application state change
When an application is being installed, it's shown in the "In Progress"
section, but when the installation is finished it had been left in
that section, even it belongs to a different section.

Similarly with the uninstall, the application was sorted to the top
of the section, but it might be better to be added to the "In Progress"
section.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1492