lib: Add a worker thread implementation
This creates a single worker thread and allows queueing of `GTask`s for
it to execute. It’s kind of like being able to call
`g_task_run_in_thread()` with a fixed thread rather than a thread pool.

This will be used in several of the plugins which require a threaded
design, to implement message passing from the plugin vfunc interface to
a single worker thread which will do the work of the plugin.

This will replace a global thread pool where multiple plugin vfuncs from
multiple plugins could be executed in parallel in transient worker
threads. This model made it hard to synchronise state, or get locking
correct, in plugins.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472