repos: Make internal hash tables immutable
Instead of modifying the internal hash tables in-place when the repos
directory changes, replace them with new hash table instances with the
updated data.

This means the hash tables can be read without a lock (but with a strong
reference) from the main thread, reducing contention on the lock to only
when the repos directory is updated.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472