repos: Load updates of the repo data asynchronously in a worker thread
Rather than loading the repo data during the first `refine()` vfunc call
after the directory has changed (and the file monitor has invalidated
it), load it in a worker thread *when* the directory changes.

This means that the refine vfunc is now non-blocking, which means it can
be implemented in the main thread. This reduces the plugin’s use of
worker threads from one for *every* refine operation, to one for every
time the repos directory is updated (which is expected to happen very
infrequently, if ever).

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472