gtk-style.css: Change margin for adaptive-symbolic icon
The margin is used to draw the icon rectangular, instead of square,
but gtk4 claims a runtime warning due to too large margin:

> Gtk GtkImage 0x889cf80 (image) reported min height -1942, but sizes must be >= 0

where the -1942 = (-999) + (-999) + 56. It's not needed to use such large
margin, especially when the code sets the image pixel size to 56, thus
split the pixel size evenly in the margin.