packagekit: Combine filtering app lists when refining
Rather than iterate through the app list multiple times to work out what
apps to send to different PackageKit calls, build multiple filtered
lists of apps in a single iteration and share some of the checks.

This should provide a small performance boost, but it’s probably not
measurable. More importantly, it cleans up the code a bit.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472