packagekit: Make refine asynchronous
This is a bit more complex than other refine functions, as there are
multiple operations which can proceed in parallel as part of the refine.

Return results only when they have all completed.

The handling of `gs_packagekit_helper_cb()` and its data will be
reworked in a subsequent commit. It currently (temporarily,
deliberately) leaks the helper struct.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472