gs-details-page: Reload the page only when the app is idle
It doesn't look good when the user installs/removes an app from the details page
and the page is reloading in the middle of the install/remove process. This can
be seen for example with flatpaks, when the applications installs also its
runtime. There are issued several reload requests, after each installed
part of the app or the runtime (the runtime, the locale part, the app itself).

The reload can lost track of the install progress too, when it gets a different
app instance during the reload.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1510