packagekit: Workaround PkControl main context assignment
The PkControl has signals for the repository changes and update changes,
which the plugin connects to, but those are never delivered, because
the main context used for the D-Bus signal delivery is not the main
context, but one pushed as a thread default from the pk-control-sync.c.
It's even worse, because the GDBusProxy under the PkControl is created
on demand, with no way to create it.