flatpak: Don't use GS_APP_STATE_AVAILABLE_LOCAL for flatpakref files
Use GS_APP_STATE_AVAILABLE instead. While the flatpakref file is
available locally, the app is available in a remote repo pointed to by
the local file. If you look at the places where the AVAILABLE_LOCAL
state is checked, they assume it doesn't make sense to check remote data
if something is locally available, but in the case of a flatpakref, a
remote will have been created and it does make sense to check it. The
only place where it is questionable whether GS_APP_STATE_AVAILABLE is
what we want for these files is in gs_details_page_refresh_all(), where
the "Software Repository Included" (infobar_details_app_repo) message is
shown for AVAILABLE_LOCAL apps. However gs_flatpak_file_to_app_ref()
adds the origin remote from the flatpakref before displaying the details
page to the user to let them choose whether to install the app, whereas
the "Software Repository Included" dialog might make the user think the
repo would be added upon installation, so it is possibly good to avoid
it.

This change affects the UI when a flatpakref file is loaded: instead of
showing "Local file (Flatpak)" as the Source in the top right corner,
with the full filename available if you click on it, we now show
"Flathub (Flatpak)" as the source, which to me seems more accurate.