gs-plugin-job: Allow it to be derived
The plan is to add a load of derived classes representing the different
kinds of job, and then to make `GsPluginJob` abstract.

This means the data specific to each job (its inputs and outputs) can be
defined and documented in a more type-safe way on each subclass of
`GsPluginJob`. It also means that the implementation of each job can be
moved out of the spaghetti of code in `GsPluginLoader` and into a vfunc
in each `GsPluginJob` subclass.

Unfortunately this means a lot of code churn to move the `GsPluginJob`
data to `GsPluginJobPrivate`. This was done as a find-and-replace of
`self->` with `priv->`, and then manually adding the declarations of
`priv` to each function. It should not need detailed review.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472