gs-app: Store a weak pointer to the GsPlugin rather than the plugin name
This removes a lot of `g_strcmp0()` calls in all the plugin functions. I
haven’t measured the performance improvement. It’s not likely to be
large, but anything is good.

More importantly, this means that the `GsPlugin` for a `GsApp` can now
be accessed without having to have a pointer to the `GsPluginLoader` to
look it up.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>