gs-plugin-loader: Run an explicit refine after ACTION_GET_POPULAR
Previously, the code converted the job from `ACTION_GET_POPULAR` into
`ACTION_REFINE` and carried on executing.

This is a bit hacky, and gets in the way of current refactoring to split
out the refine job handling into a separate class. Change it to run an
explicit refine job on the popular list instead.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472