gs-plugin: Add a setup vfunc
This will allow plugins to be set up asynchronously.

The previous `gs_plugin_setup()` function was synchronous and blocking,
which didn’t allow the plugins to be set up in parallel.

No plugins currently implement this vfunc, but the plugins which
currently implement `gs_plugin_setup()` will be ported to it in the
following commits.

`GsPluginLoader` has not yet been ported to use this vfunc instead of
`gs_plugin_setup()`; that will happen in a following commit. The
intermediate state of the source tree before then may be non-functional.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472