Revive webapp support
This is a re-implementation of the epiphany plugin (dropped in
ea5eb222f) in a way that should be more robust and maintainable.

This is the first step toward bringing back webapp support. I am doing
this as a project sponsored by a GNOME Foundation grant. The goal is to
include a set of Progressive Web Apps in Software that can be installed
via Epiphany's web app support (and potentially also via Chromium), to
expand the selection of apps available to users who might not otherwise
find them. While Epiphany's web apps do not currently implement support
for PWA manifests, that is hopefully going to be in scope for this
project.  While the plan is to only include PWAs in the set hard coded
into Software, non-PWA web apps installed via Epiphany will also show up
in Software.

The previous implementation of web apps in Software was dropped because
it was buggy and users did not like it, since they didn't perceive much
benefit of using sites as web apps versus using a normal browser and
were confused by the apps not being native desktop apps. The following
factors should mitigate or address those issues for the new
implementation:
1) We're going to use a D-Bus API provided by Epiphany for enumerating,
   installing, and removing web apps rather than re-implementing those
   functions in Software. This avoids bugs that can occur when the
   implementations are out of sync.
2) We're going to differentiate web apps from native apps in the UI,
   pending design input on how to do so.
3) The set of PWAs included with Software will be mostly or entirely
   apps that would not otherwise be available to the user, because they
   are only available as PWAs and not native apps.
4) Once we have support for PWA manifests in Epiphany, or support for
   Chromium-based web apps which already support manifests, the apps
   should be more featureful and closer to native apps than the current
   web app implementation, e.g. they may work offline to some extent.

(This removes the dependency on epiphany-runtime in the spec file,
because it's not clear yet if this feature will be enabled by default in
Fedora and the plan is to make this work with flatpak'd Epiphany which
doesn't have an analogous way to install epiphany without a desktop
icon; see https://bugzilla.redhat.com/show_bug.cgi?id=1781359 for
context.)

The file plugins/epiphany/org.gnome.Epiphany.WebAppProvider.xml comes
from its canonical source:
https://gitlab.gnome.org/GNOME/epiphany/-/blob/master/src/webapp-provider/org.gnome.Epiphany.WebAppProvider.xml