gs-flatpak: Ensure installed GsApps are marked as such at creation time
It seems it’s possible for a `GsApp` to be in the cache when
`create_installed()` is called, with a state of `AVAILABLE`. That
prevents the app from being marked as `INSTALLED`, even though
`create_installed()` is only called from places where we know the app is
installed.

This points to some kind of missing cache invalidation/update when an
app goes from available → installed, but I have not managed to track it
down.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>