tests: Skip parental controls in flatpak
If the flatpak in the test environment is built with parental controls
support, the tests will fail when the dbus system bus isn't available.
Set the `FLATPAK_SYSTEM_HELPER_ON_SESSION` environment variable so
flatpak skips parental controls. This is the same thing flatpak's test
suite does in lieu of proper malcontent mocking or some non-dbus method
of being disabled.