packagekit: Rework progress callback data handling
The progress callback data closure, `GsPackagekitHelper`, exists to make
three things available to the progress callback
(`gs_packagekit_helper_cb()`):
 1. The `GsPlugin`, so its status can be updated.
 2. An optional `GsApp`, so its progress can be updated.
 3. An optional `GsAppList`, so the `GsApp` can be looked up or,
    failing that, the progress of the entire list can be updated.

Eventually, it should be possible to remove the `GsPlugin` data, as it
makes more sense to report progress on a job-by-job basis than a
plugin-by-plugin basis (what does the ‘progress’ of a plugin mean unless
it’s associated with an ongoing operation?). At that point, the closure
could be slimmed down to just a `GsAppList`, as that will contain all
the right `GsApp`s already.

Until that point, the closure has to be kept, and has to be kept alive
for zero or more calls between the `async()` operation being started,
and the `GAsyncReadyCallback` being called on its completion.

The best way to do that is to keep a list of progress closures to free
when the overall operation closure (`RefineData`) is freed.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472