tests: Use correct result lists when checking wildcard resolves
Since the introduction of `GsPluginJobRefine`, the original `GsAppList`
passed as an argument to the job is no longer modified, and the results
are returned as a separate list. This move towards treating lists as
more immutable should make it easier to keep things threadsafe in
future.

`gs_plugin_loader_job_process_finish()` already handles that and returns
the results list, so most code doesn’t need changes. However, these
tests are using a wrapper which doesn’t return a result list at all. So
they need modifying to explicitly check the correct list.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>