gs-os-update-page: Mark installed apps for removal
When an app exists in the GsOSUpdatePage, it should have been
created by a plugin with the goal of adding, removing,
updating or downgrading it. Therefore, if the app is in the
GS_APP_STATE_INSTALLED, it should be safe for GS to assume that
the app has to be removed.

An example situation where this is useful is when an package
upgrade removes a dependency on an OS package and the "dependent"
package can then be safely removed. This might not be a big
matter in stable distros, but is definitely something that happens
often in rolling-release channels.