flatpak: Set no-interaction correctly when refreshing
A GS_PLUGIN_FLAGS_INTERACTIVE flags will be set based on whether the
plugin action should be interactive or not. Set no-interaction value
of FlatpakInstallation according to that flag.

This fixes the issue where an polkit authentication dialog pops up
when GNOME Software is running in the background and trying to check
updates, if the polkit rules for
org.freedesktop.Flatpak.appstream-update requires authorisation.

https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1561