packagekit: Merge packagekit-refine-repos plugin into the main plugin
There was no reason to keep it separate, and since commit 83f5653b, the
packagekit-refine-repos plugin was no longer correctly matching with the
packagekit plugin as the chosen management plugin of PackageKit apps and
repos.

Handily, the existing `search_files_cb()` callback in the packagekit
plugin does exactly what we need for repos as well as apps, which means
that most of the code in the packagekit-refine-repos plugin can be
dropped as it duplicates code already present in the packagekit plugin.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472