repos: Drop plugin ordering requirements
The app IDs are set from construction time, so there’s no need for repos
refine to happen after packagekit or rpm-ostree refine.

This fixes a cycle in plugin construction introduced when merging the
packagekit-refine-repos plugin into packagekit.

Ultimately it would be nice to be able to get rid of the repos plugin,
and instead have the packagekit and rpm-ostree plugins set the
appropriate data on their apps and repos themselves. Doing it in a
separate plugin leads to ordering problems, increased latency, and
problems with keeping data in sync between the plugins, system, apps and
repo objects.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472