packagekit: Merge systemd-updates plugin into the main plugin
There was no reason to keep it separate, and since commit 83f5653b, the
systemd-updates plugin was no longer correctly matching with the
packagekit plugin as the chosen management plugin of PackageKit apps and
repos.

Rather than clearing the download size of apps to zero as part of a
second pass in `refine_async()` (if they’re
already prepared for an offline update), modify the code in
`packagekit-common.c` which sets the download size in the first place,
to correctly account for prepared updates.

This will probably fix offline update download size handling in other
situations — this code is called by `gs_plugin_url_to_app()`. That code
didn’t previously account for prepared offline updates.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472