appstream: appstream files from user-data dirs when appropriate
gs_external_appstream_refresh_user installs external
appstream files in non-standard XDG_DATA_HOME directory.
Make sure that the corresponding xml and yaml directories
are read so that the installed files do not get ignored.

Additionally, changes to "external-appstream-system-wide" can leave old
files in the directories where the files are installed. Make sure we do not
read those undesired files. When the key is TRUE, it is enough to
ignore the custom location where user files are installed, as it is
a non-standard location only used by GNOME Software. However, when
the key is FALSE, ignoring the system directory where the system
files are installed is a mistake, as /var/cache/app-info/xmls is
a standard directory and other programs might install files there.
In this case, match also the prefix to make sure we are only ignoring
our own dirty files