gs-shell: Show the window before activating it
The gs_shell_activate() only called gtk_window_present(), which just
shows the window when it's hidden, which is not enough to gain the focus
when the Software is activated with the main window hidden. Showing
the window beforehand makes sure the gtk_window_present() also focuses
the window. Without this the window can be shown, but stay below
the other windows.

Closes https://gitlab.gnome.org/GNOME/gnome-software/-/issues/1572