flatpak: Store interactive state on transactions instead of plugin
As much as is possible at the moment, store the interactive state on
each transaction, rather than always querying the plugin to see if it’s
interactive.

This should reduce the chance that long-running flatpak operations which
start as non-interactive later become interactive as the flatpak plugin
as a whole has since become interactive (or vice-versa).

It’s one step towards eventually removing `GS_PLUGIN_FLAGS_INTERACTIVE`,
which promotes writing code which is subject to that kind of bug.

This commit uses the new `flatpak_transaction_get_no_interaction()` API,
added in flatpak 1.13.0. If that’s not available, it falls back to a
workaround using GObject data.

The `flatpak_transaction_set_no_interaction()` API has existed in
libflatpak for a while, which is why the conditional compilation for
the `{get,set}_no_interaction()` pair is a little complex and has to use
`#define`s rather than `static inline`s.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472