show-metainfo: Set the application also to the featured carousel on the Overview page
By using the --show-metainfo a user could see how it looks in the details
page, but it was not easy to see it in the featured carousel, which this
change tries to address. The app is shown in the Details page and when
the back button is pressed it's also visible in the carousel.