gs-plugin-event: Split finalize and dispose vfuncs
`dispose()` is needed to be able to clear refcounting loops. While I
haven’t seen any of them, it’s possible that at some point a refcounting
loop might exist between an app and an event, so split the vfuncs in
case.

Also tidy up the code to use ‘clear’ functions.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>