gs-plugin: Split out and refactor download code
Split out the bulk of the download code and rewrite it to be
asynchronous. Keep a wrapper implementation in `gs-plugin.h` which
binds the new implementation to the existing `gs_plugin_download_file()`
and `gs_plugin_download_rewrite_resource()` APIs, until all callers of
those APIs can be ported.

This rewrite of the download code brings a few advantages:
 * Local and remote code paths are combined so they both get more
   testing.
 * The download is chunked so that the entire downloaded file isn’t
   allocated in memory near the end of the download. This significantly
   reduces the heap allocation sizes.
 * The code operates as a splice, so downloading to different kinds of
   `GOutputStream` is supported in the same code path. While
   `gs_plugin_download_data()` wasn’t actually used any more (and hence
   has been deleted), if it needs to be re-added in future, this can
   easily be done using a `GMemoryOutputStream`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472