gs-download-utils: Split out SoupSession construction
Centralising it in `GsPlugin` and `GsPluginLoader` is not great, since
it means connection and authentication information get shared between
different plugin contexts (which don’t need to share them), and thread
contention on the session is increased.

We probably want to aim for a model where, roughly, there is one
`SoupSession` per plugin which needs one.

Factor out a common construction method for `SoupSession` which sets the
gnome-software user agent and default properties. Plugins can then call
this.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472