fwupd: Remove unhelpful helper function
If refreshing metadata in the fwupd plugin is to be async,
`gs_plugin_fwupd_refresh_remote()` would have to be made async. That’s
quite a lot of boilerplate for very little function code — so it seems
to make more sense to split the function into a helper to check the
cache age, and the actual libfwupd call. The libfwupd call already has
an async version.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472