data: Drop upgrade-background-uri GSettings key
The background image for upgrades should not be settable per-user. This
was a hack to allow the background image to be set differently for
different distributions using GSettings overrides.

That’s now possible by dropping the background image in a well-known
path, as documented in `doc/vendor-customisations.md`, so this hack can
be removed.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>