doc: Document featured and popular applications for vendors
Document the current state of how featured and popular applications are
chosen, and what that means for the UI.

There are several different bits of metadata taken into account, and it
would probably make sense to rationalise them at some point. I have made
no attempt to do that here, though; this commit is purely to document
the current state of the system.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>