fedora-pkgdb-collections: Split JSON parsing out into a helper
It has quite complex control flow, so if `_ensure_cache()` is to be
made asynchronous, that will require a lot of `g_task_return_boolean()`
branches.

Easier to split it out into a helper function and call that, given that
the parsing is synchronous.

This will help with future refactoring.

This introduces no functional changes.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472