Remove parent instance from Private structs
A number of types were switched from G_DECLARE_FINAL_TYPE to
G_DECLARE_DERIVABLE_TYPE, and their structs renamed from GsFoo to
GsFooPrivate, but the first element, embedding the parent type, was not
removed.

This is unneeded in the Private structure. G_DECLARE_DERIVABLE_TYPE
defines an empty struct wrapping the parent.