gs-plugin-loader: Factor out gs_plugin_loader_hint_job_finished()
Some plugin jobs might cause several signal emissions of changes as they
progress. The plugin loader currently queues those up and emits them
when a job is finished.

If jobs are being moved out to external `GsPluginJob` objects, those
objects need a way to signal to the `GsPluginLoader` that they are done
potentially emitting signals.

Refactor the existing code to do that, and expose it in a
`GsPluginLoader` API.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472