gs-utils: Let gs_utils_get_file_age() return guint64
Since https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/1219 the cache
file age is turned into guint64. Having the gs_utils_get_file_age() return
guint and the cache file not existing meant returning G_MAXUINT, which is smaller
than G_MAXUINT64, thus when the file did not exist it was behaving like
the file exists and it failed later in the code.

This could be seen on start as a runtime warning when there's no ratings.json file
in the cache and automatic updates are enabled.