gs-odrs-provider: Make refreshing asynchronous
Change the refresh API to be asynchronous, and to no longer rely on a
`GsPlugin` being passed in to provide progress updates.

This will allow `GsOdrsProvider` to be used in contexts where a
`GsPlugin` isn’t available, and it decouples progress updates from the
locking semantics of setting properties on a `GsApp`.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1472