gs-feature-tile: Correct narrow mode switch on width change
The GtkButton::size_allocate() is not called at all, thus the narrow
mode switch did not work. The way to do it in gtk4 is to use a GtkLayoutManager.
Unfortunately, the GtkBinLayout is not a derivable type, thus this copies
it into the sources and adds the narrow mode code into the allocate()
method.