gs-plugin-job-refine: Don’t recursive if refine flags are empty
The refine job recurses into related components of the components it’s
been asked to refine, also refining them. For example, addons and
runtimes of the apps it’s refining.

To avoid infinite recursion, these recursive refines are done with
various flags disabled. If those are the only flags in the set of
`GsPluginRefineFlags`, though, the recursive refine operations will be
no-ops.

Might as well not do the recursion in those cases.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #1658